split inbound control status composer
